perf: optimize the generation of targets under src
