+++ src/exe/include/helpers.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <spdlog/spdlog.h>
+
+namespace exe {
+namespace helpers {
+inline int some_fun() {
+    spdlog::info("Hello helpers!");
+    return 0;
+}
+} // namespace helpers
+} // namespace exe
