+++ src/header/include/header/header.hpp
@@ -1,0 +1,18 @@
+#pragma once
+
+#include <spdlog/spdlog.h>
+#include "git.h"
+
+namespace header {
+namespace common {
+inline int some_fun() {
+    spdlog::info("{}", git::ProjectVersion());
+    return 0;
+}
+
+const std::string_view const_string() {
+    spdlog::info("Calling {}", __FUNCTION__);
+    return "This is a header_only only const string";
+}
+} // namespace common
+} // namespace header
