--- src/header_only/CMakeLists.txt
@@ -1,31 +1,0 @@
-set(target_name header_only)
-
-find_package(spdlog CONFIG REQUIRED)
-
-add_library(${target_name} INTERFACE)
-
-target_include_interface_directories(
-  ${target_name} ${CMAKE_CURRENT_SOURCE_DIR}/include
-  ${CMAKE_BINARY_DIR}/git/include)
-
-target_link_libraries(${target_name} INTERFACE spdlog::spdlog)
-
-install_target(
-  NAME
-  ${target_name}
-  VERSION
-  ${CMAKE_PROJECT_VERSION}
-  TARGETS
-  ${target_name}
-  INCLUDES
-  ${CMAKE_CURRENT_SOURCE_DIR}/include/${target_name}/
-  ${CMAKE_BINARY_DIR}/git/include/
-  LICENSE_FILE_LIST
-  ${CMAKE_SOURCE_DIR}/LICENSE
-  DEPENDENCIES
-  spdlog::spdlog)
-
-# Test
-if(BUILD_TESTING)
-  add_subdirectory(tests)
-endif()
