+++ template/[% if docs_type == 'sphinx' %]docs[% endif %]/conf.py.jinja
@@ -54,8 +54,8 @@ primary_domain = "cpp"
 highlight_language = "cpp"
 
 cpp_id_attributes = [
-[%- if add_compile_target == true %]
-    "COMPILE_EXPORT",
+[%- if compile_target != '' %]
+    "{{ compile_target|upper }}_EXPORT",
 [%- endif %]
     "GIT_EXTERN_C_BEGIN",
 ]
