--- template/src/CMakeLists.txt.jinja
@@ -1,12 +1,0 @@
-[%- if add_compile_target == true -%]
-# A compile library
-add_subdirectory(compile)
-[%- endif %]
-[%- if add_executable_target == true %]
-# A executable
-add_subdirectory(executable)
-[%- endif %]
-[%- if add_header_only_target == true %]
-# A header-only library
-add_subdirectory(header_only)
-[%- endif %]
