+++ template/src/[% if compile_target != '' %]{{ compile_target }}[% endif %]/CMakeLists.txt.jinja
@@ -1,4 +1,4 @@
-set(target_name compile)
+set(target_name {{ compile_target }})
 
 find_package(spdlog CONFIG REQUIRED)
