+++ template/src/[% if compile_target != '' %]{{ compile_target }}[% endif %]/include/{{ compile_target }}.hpp.jinja
@@ -1,0 +1,19 @@
+#pragma once
+
+#include <string>
+#include "compile_export.hpp"
+
+namespace {{ compile_target }} {
+namespace info {
+/**
+ * \brief Get {{ compile_target }} version string
+ * \return The distribution {{ compile_target }} version provided.
+ */
+std::string_view COMPILE_EXPORT version_compile() noexcept;
+
+/**
+ * \return A bool represents if it is debug distribution.
+ */
+bool COMPILE_EXPORT is_debug() noexcept;
+}; // namespace info
+}; // namespace {{ compile_target }}
