+++ template/src/[% if compile_target != '' %]{{ compile_target }}[% endif %]/src/{{ compile_target }}.cpp.jinja
@@ -1,0 +1,20 @@
+#include "{{ compile_target }}.hpp"
+
+#include "git.h"
+
+namespace {{ compile_target }} {
+namespace info {
+std::string_view version_compile() noexcept {
+    return git::ProjectVersion();
+}
+
+bool is_debug() noexcept {
+#ifdef _DEBUG
+    return true;
+#else
+    return false;
+#endif
+}
+
+} // namespace info
+} // namespace {{ compile_target }}
