+++ template/src/[% if compile_target != '' %]{{ compile_target }}[% endif %]/tests/test_{{ compile_target }}.cpp.jinja
@@ -1,0 +1,16 @@
+#include <gtest/gtest.h>
+#include "{{ compile_target }}.hpp"
+
+TEST(info, version) {
+    const auto version = {{ compile_target }}::info::version_compile();
+    EXPECT_FALSE(version.empty());
+}
+
+TEST(info, distribution) {
+    const auto is_debug = {{ compile_target }}::info::is_debug();
+#ifdef _DEBUG
+    GTEST_ASSERT_TRUE(is_debug);
+#else
+    GTEST_ASSERT_FALSE(is_debug);
+#endif
+}
