+++ template/src/[% if exe_target != '' %]{{ exe_target }}[% endif %]/include/helpers.hpp.jinja
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <spdlog/spdlog.h>
+
+namespace {{ exe_target }} {
+namespace helpers {
+inline int some_fun() {
+    spdlog::info("Hello helpers!");
+    return 0;
+}
+} // namespace helpers
+} // namespace {{ exe_target }}
