+++ template/src/[% if exe_target != '' %]{{ exe_target }}[% endif %]/src/main.cpp.jinja
@@ -1,0 +1,9 @@
+﻿#include <spdlog/spdlog.h>
+
+#include "git.h"
+#include "helpers.hpp"
+
+int main() {
+    spdlog::info("Get a returned value: {} ; Version: {}", {{ exe_target }}::helpers::some_fun(), git::ProjectVersion());
+    return 0;
+}
