+++ template/src/[% if exe_target != '' %]{{ exe_target }}[% endif %]/tests/test_{{ exe_target }}_helper.cpp.jinja
@@ -1,0 +1,7 @@
+#include <gtest/gtest.h>
+#include "helpers.hpp"
+
+TEST({{ exe_target }}_helpers, some_fun) {
+    const auto fun_ret = {{ exe_target }}::helpers::some_fun();
+    GTEST_ASSERT_EQ(0, fun_ret);
+}
