+++ template/src/[% if header_target != '' %]{{ header_target  }}[% endif %]/tests/test_{{ header_target }}.cpp.jinja
@@ -1,0 +1,12 @@
+#include <gtest/gtest.h>
+#include "{{ header_target }}/{{ header_target }}.hpp"
+
+TEST(common, some_fun) {
+    const auto fun_ret = {{ header_target }}::common::some_fun();
+    GTEST_ASSERT_EQ(0, fun_ret);
+}
+
+TEST(common, const_string) {
+    const auto fun_ret = {{ header_target }}::common::const_string();
+    GTEST_ASSERT_GT(fun_ret.size(), 0);
+}
